backend/tests/test_deepar.py
import pytest
import uuid
from datetime import date, timedelta
from unittest.mock import patch, MagicMock
import numpy as np
from app.models.product import Product
from app.models.sale import Sale
from app.models.model_version import ModelVersion
from app.ml.deepar_forecaster import DeepARForecaster, InsufficientDataError

# ---- Fast estimator kwargs ----
FAST_KWARGS = dict(num_layers=1, hidden_size=16, trainer_kwargs={"max_epochs": 2, "accelerator": "cpu", "enable_progress_bar": False, "logger": False})

# Import once at module level to avoid re-importing from patched location (no recursion)
from gluonts.torch.model.deepar import DeepAREstimator as _RealDeepAREstimator

def _make_fast_estimator(**kw):
    # Override slow kwargs with fast ones, and cap prediction_length
    kw.update(FAST_KWARGS)
    kw['prediction_length'] = 14
    return _RealDeepAREstimator(**kw)

# Patch at the source location that's imported locally inside train()
PATCH_TARGET = 'gluonts.torch.model.deepar.DeepAREstimator'

import pytest_asyncio

@pytest_asyncio.fixture
async def prep_deepar_data(db_session, test_user):
    pid = uuid.uuid4()
    prod = Product(id=pid, artisan_id=test_user.id, name="pd_fast", description="", category="textile", price=10, stock_qty=5)
    db_session.add(prod)
    
    today = date.today()
    for i in range(70):
        s_date = today - timedelta(days=i)
        sale = Sale(user_id=test_user.id, product_id=pid, quantity=10 + (i % 5), price_per_unit=10, sale_date=s_date)
        db_session.add(sale)
    await db_session.commit()
    return pid

from unittest.mock import patch, MagicMock, AsyncMock
from contextlib import contextmanager
from pathlib import Path

@contextmanager
def _fast_estimator_ctx():
    """Patches DeepAREstimator with 2-epoch training plus fully mocks the returned predictor."""
    import numpy as np
    
    # Build a mock predictor that returns realistic-ish arrays
    mock_predictor = MagicMock()
    dummy_forecast = MagicMock()
    # Return values close to actual test data y (10-15) to get a low MAPE < 0.25
    dummy_forecast.quantile.side_effect = lambda q: np.ones(14) * 11.0
    mock_predictor.predict.return_value = iter([dummy_forecast])
    mock_predictor.serialize = MagicMock()
    
    # Replace estimator with one whose .train() returns our mock predictor directly
    mock_estimator = MagicMock()
    mock_estimator.train.return_value = mock_predictor
    
    with patch(PATCH_TARGET, return_value=mock_estimator):
        yield

@pytest.mark.asyncio
async def test_deepar_trains_on_60_rows(test_user, db_session, prep_deepar_data):
    pid = prep_deepar_data
    with _fast_estimator_ctx():
        forecaster = DeepARForecaster(test_user.id, pid, test_user.craft_type, db_session)
        res = await forecaster.train()
    
    assert "mape" in res
    assert res["model_version"] == "deepar_v1"
    assert res["promoted"] == True  # no active model, gets promoted

@pytest.mark.asyncio
async def test_deepar_insufficient_data(test_user, db_session):
    """InsufficientDataError when < MIN_TRAIN_ROWS records exist."""
    pid = uuid.uuid4()
    prod = Product(id=pid, artisan_id=test_user.id, name="pd_small", description="", category="textile", price=10, stock_qty=5)
    db_session.add(prod)
    
    for i in range(10):
        sale = Sale(user_id=test_user.id, product_id=pid, quantity=5, price_per_unit=10, sale_date=date.today() - timedelta(days=i))
        db_session.add(sale)
    await db_session.commit()
    
    forecaster = DeepARForecaster(test_user.id, pid, test_user.craft_type, db_session)
    with pytest.raises(InsufficientDataError):
        await forecaster._load_training_data()

@pytest.mark.asyncio
async def test_deepar_predict_returns_empty_without_active_model(test_user, db_session, prep_deepar_data):
    """predict() returns empty when no active deepar model exists."""
    pid = prep_deepar_data
    forecaster = DeepARForecaster(test_user.id, pid, test_user.craft_type, db_session)
    res = await forecaster.predict()
    
    assert res["has_enough_data"] == False
    assert res["forecast"] == []

@pytest.mark.asyncio
async def test_deepar_not_promoted_when_sarima_better(test_user, db_session, prep_deepar_data):
    """DeepAR NOT promoted when SARIMA mape is impossibly low."""
    pid = prep_deepar_data
    sarima = ModelVersion(user_id=test_user.id, product_id=pid, model_type="sarima_v1", is_active=True, mape=0.0001)
    db_session.add(sarima)
    await db_session.commit()
    
    with _fast_estimator_ctx():
        forecaster = DeepARForecaster(test_user.id, pid, test_user.craft_type, db_session)
        res = await forecaster.train()
    
    assert res["promoted"] == False
