backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
httpx
pytest
pytest-asyncio
alembic
email-validator
groq
statsmodels>=0.14.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
apscheduler>=3.10.0
pytrends>=4.9.0
